Adding the HTTP service
