+++ packages/realm-web-integration-tests/my-test-app-template/services/custom-http-service/config.json
@@ -1,0 +1,6 @@
+{
+    "name": "custom-http-service",
+    "type": "http",
+    "config": {},
+    "version": 1
+}
