+++ packages/realm-web-integration-tests/my-test-app-template/services/custom-http-service/rules/allow-everything.json
@@ -1,0 +1,12 @@
+{
+    "id": "5e8683fdb739bee989c7ef8e",
+    "name": "allow-everything",
+    "actions": [
+        "get",
+        "post",
+        "put",
+        "delete",
+        "patch",
+        "head"
+    ]
+}
